silence TS errors in tests (WIP)
