+++ packages/nitro-protocol/test/contracts/MultiAssetHolder/transfer.test.ts
@@ -1,7 +1,6 @@
 import {expectRevert} from '@statechannels/devtools';
 import {Contract, BigNumber} from 'ethers';
 
-import AssetHolderArtifact from '../../../artifacts/contracts/test/TESTAssetHolder.sol/TESTAssetHolder.json';
 import {
   allocationToParams,
   getRandomNonce,
@@ -9,11 +8,8 @@ import {
   randomChannelId,
   randomExternalDestination,
   replaceAddressesAndBigNumberify,
-  setupContract,
 } from '../../test-helpers';
 import {encodeAllocation} from '../../../src/contract/outcome';
-
-const provider = getTestProvider();
 
 let AssetHolder: Contract;
 
@@ -26,10 +22,6 @@ const addresses = {
   A: randomExternalDestination(),
   B: randomExternalDestination(),
 };
-
-beforeAll(async () => {
-  AssetHolder = setupContract(provider, AssetHolderArtifact, process.env.TEST_ASSET_HOLDER_ADDRESS);
-});
 
 const reason0 = 'h(allocation)!=assetOutcomeHash';
 const reason1 = 'Indices must be sorted';
